app/requirements.txt
Flask
gunicorn
python-decouple
requests
